+++ web/src/App.tsx
@@ -7,8 +7,8 @@ import ProtectedRoute from '@/components/ProtectedRoute';
 import LoginPage from '@/pages/login/LoginPage';
 import RegisterPage from '@/pages/RegisterPage';
 import EventsPage from '@/pages/EventsPage';
-import EventDetailPage from '@/pages/EventDetailPage';
 import CompetePage from '@/pages/CompetePage/CompetePage';
+import EventDetailPage from '@/pages/EventDetail/EventDetailPage';
 import ProfilePage from '@/pages/ProfilePage';
 import TrainingPage from '@/pages/TrainingPage';
 import TrainingLeaderboardPage from '@/pages/TrainingLeaderboardPage';
