+++ web/src/pages/EventDetail/Header/EventStatus/ActiveEvent.tsx
@@ -1,0 +1,11 @@
+export default function ActiveEvent() {
+  return (
+    <span className="flex items-center gap-1.5 px-2 py-0.5 rounded-full bg-emerald-500/10 border border-emerald-500/20">
+      <span className="relative flex h-2 w-2">
+        <span className="animate-ping absolute inline-flex h-full w-full rounded-full bg-emerald-400 opacity-75"></span>
+        <span className="relative inline-flex rounded-full h-2 w-2 bg-emerald-500"></span>
+      </span>
+      <span className="text-xs text-emerald-500 font-medium">LIVE</span>
+    </span>
+  )
+}
