+++ web/src/pages/EventDetail/Tabs/Overview/OverviewDescription.tsx
@@ -1,0 +1,15 @@
+import Markdown from "@/components/Markdown";
+
+export default function OverviewDescription({description}: { description: string | undefined }) {
+  return (
+    <>
+      {description ? (
+        <div className="prose prose-sm prose-invert max-w-none">
+          <Markdown content={description} />
+        </div>
+      ) : (
+        <p className="text-muted-foreground text-sm">No description provided.</p>
+      )}
+    </>
+  )
+}
